fix vercel lint error
